Angle problem solved, and partially implemented!
Orbits worked only for y=0 and any x value and only for a positive y velocity. It turns out that the angle parameter requires adding the starting angle to correct the 'error' experienced in previous pull requests.

The final part to the angle problem comes in the form of an unquestioned presumption. The angular_progression variable in the body object is incrementing up and, in the code, is interpreted as a clockwise orbit. It however does not work for anti-clockwise orbits, and I will have to fix this before moving onto the next stage.

Also removed outdated code in body.h functions.